fix(admin): keep retention backfill script scoped
