+++ scripts/backfill_retention_metrics.ts
@@ -16,7 +16,6 @@
  *   bun run stripe:backfill-retention-metrics --events-file=./tmp/stripe-events.json --from=2026-01-01 --to=2026-04-23
  */
 import type Stripe from 'stripe'
-import type { RevenueMovement, RevenuePlanRow, StripeInfoRevenueState } from '../supabase/functions/_backend/utils/revenue_metrics.ts'
 import type { Database } from '../supabase/functions/_backend/utils/supabase.types.ts'
 import { existsSync } from 'node:fs'
 import { mkdir, readFile, writeFile } from 'node:fs/promises'
@@ -24,7 +23,6 @@ import process from 'node:process'
 import { createClient } from '@supabase/supabase-js'
 import { Client as PgClient } from 'pg'
 import StripeClient from 'stripe'
-import { calculateChurnRevenue, calculateNrr, classifyRevenueMovement, getEventDateId, getPreviousDateId, getSubscriptionMrr, hasRevenueMovement } from '../supabase/functions/_backend/utils/revenue_metrics.ts'
 
 const DEFAULT_ENV_FILE = './internal/cloudflare/.env.prod'
 const DEFAULT_LOOKBACK_DAYS = 30
@@ -41,8 +39,17 @@ const SUBSCRIPTION_EVENT_TYPES = [
 type SupabaseClient = ReturnType<typeof createClient<Database>>
 type DailyRevenueMetricRow = Database['public']['Tables']['daily_revenue_metrics']['Row']
 type DailyRevenueMetricInsert = Database['public']['Tables']['daily_revenue_metrics']['Insert']
+type PlanRow = Database['public']['Tables']['plans']['Row']
 type StripeStatus = Database['public']['Enums']['stripe_status']
 type SubscriptionEventType = typeof SUBSCRIPTION_EVENT_TYPES[number]
+type StripeInfoRevenueState = {
+  is_good_plan?: boolean | null
+  paid_at?: string | null
+  price_id?: string | null
+  product_id?: string | null
+  status?: StripeStatus | null
+} | null | undefined
+type RevenuePlanRow = Pick<PlanRow, 'price_m' | 'price_m_id' | 'price_y' | 'price_y_id' | 'stripe_id'>
 
 interface CustomerRevenueBaselineRow {
   customer_id: string
@@ -130,6 +137,30 @@ interface RetentionMetricSummaryRow {
   retained_churn_mrr: number | string | null
   retained_contraction_mrr: number | string | null
   retained_expansion_mrr: number | string | null
+}
+
+interface RevenueMovement {
+  currentMrr: number
+  nextMrr: number
+  newBusinessMrr: number
+  expansionMrr: number
+  contractionMrr: number
+  churnMrr: number
+}
+
+interface DailyRevenueChangeSummary {
+  churnMrr: number
+  contractionMrr: number
+  expansionMrr: number
+}
+
+const ZERO_REVENUE_MOVEMENT: RevenueMovement = {
+  currentMrr: 0,
+  nextMrr: 0,
+  newBusinessMrr: 0,
+  expansionMrr: 0,
+  contractionMrr: 0,
+  churnMrr: 0,
 }
 
 function getArgValue(args: string[], prefix: string): string | null {
@@ -419,6 +450,133 @@ function getPreviousSubscriptionStatus(event: Stripe.Event) {
     hasStatus: true,
     status: toBackfillStripeStatus(previousAttributes.status),
   }
+}
+
+function getRevenueMetricDateId(targetDate = new Date()) {
+  return new Date(Date.UTC(targetDate.getUTCFullYear(), targetDate.getUTCMonth(), targetDate.getUTCDate())).toISOString().slice(0, 10)
+}
+
+function getEventDateId(eventOccurredAtIso: string) {
+  return new Date(eventOccurredAtIso).toISOString().slice(0, 10)
+}
+
+function getPreviousDateId(dateId: string) {
+  const target = new Date(`${dateId}T00:00:00.000Z`)
+  target.setUTCDate(target.getUTCDate() - 1)
+  return getRevenueMetricDateId(target)
+}
+
+function getPlanMrr(plan: RevenuePlanRow | null | undefined, priceId: string | null | undefined) {
+  if (!plan || !priceId)
+    return 0
+
+  if (plan.price_m_id === priceId)
+    return Number(plan.price_m) || 0
+
+  if (plan.price_y_id === priceId)
+    return (Number(plan.price_y) || 0) / 12
+
+  return 0
+}
+
+function getPlanByProductId(plans: RevenuePlanRow[], productId: string | null | undefined) {
+  if (!productId)
+    return null
+
+  return plans.find(plan => plan.stripe_id === productId) ?? null
+}
+
+function getSubscriptionMrr(plans: RevenuePlanRow[], stripeInfo: StripeInfoRevenueState) {
+  if (stripeInfo?.status !== 'succeeded' || stripeInfo?.is_good_plan === false)
+    return 0
+
+  return getPlanMrr(getPlanByProductId(plans, stripeInfo.product_id), stripeInfo.price_id)
+}
+
+function classifyRevenueMovement(
+  currentStripeInfo: StripeInfoRevenueState,
+  nextStripeInfo: StripeInfoRevenueState,
+  plans: RevenuePlanRow[],
+): RevenueMovement {
+  const currentMrr = getSubscriptionMrr(plans, currentStripeInfo)
+  const nextMrr = getSubscriptionMrr(plans, nextStripeInfo)
+
+  if (currentMrr === 0 && nextMrr === 0)
+    return { ...ZERO_REVENUE_MOVEMENT }
+
+  if (currentMrr === 0 && nextMrr > 0) {
+    if (!currentStripeInfo?.paid_at) {
+      return {
+        ...ZERO_REVENUE_MOVEMENT,
+        currentMrr,
+        nextMrr,
+        newBusinessMrr: nextMrr,
+      }
+    }
+
+    return {
+      ...ZERO_REVENUE_MOVEMENT,
+      currentMrr,
+      nextMrr,
+      expansionMrr: nextMrr,
+    }
+  }
+
+  if (currentMrr > 0 && nextMrr === 0) {
+    return {
+      ...ZERO_REVENUE_MOVEMENT,
+      currentMrr,
+      nextMrr,
+      churnMrr: currentMrr,
+    }
+  }
+
+  if (nextMrr > currentMrr) {
+    return {
+      ...ZERO_REVENUE_MOVEMENT,
+      currentMrr,
+      nextMrr,
+      expansionMrr: nextMrr - currentMrr,
+    }
+  }
+
+  if (currentMrr > nextMrr) {
+    return {
+      ...ZERO_REVENUE_MOVEMENT,
+      currentMrr,
+      nextMrr,
+      contractionMrr: currentMrr - nextMrr,
+    }
+  }
+
+  return {
+    ...ZERO_REVENUE_MOVEMENT,
+    currentMrr,
+    nextMrr,
+  }
+}
+
+function hasRevenueMovement(movement: RevenueMovement) {
+  return movement.newBusinessMrr > 0
+    || movement.expansionMrr > 0
+    || movement.contractionMrr > 0
+    || movement.churnMrr > 0
+}
+
+function calculateNrr(previousMrr: number, dailyChanges: DailyRevenueChangeSummary) {
+  if (previousMrr <= 0)
+    return 100
+
+  const retainedMrr = Math.max(
+    previousMrr - dailyChanges.churnMrr - dailyChanges.contractionMrr + dailyChanges.expansionMrr,
+    0,
+  )
+
+  return Number(((retainedMrr / previousMrr) * 100).toFixed(2))
+}
+
+function calculateChurnRevenue(dailyChanges: DailyRevenueChangeSummary) {
+  return Number((dailyChanges.churnMrr + dailyChanges.contractionMrr).toFixed(2))
 }
 
 function toRevenueState(state: TrackedSubscriptionState): NonNullable<StripeInfoRevenueState> {
